+++ hail/python/hailtop/hailctl/__init__.py
@@ -1,5 +1,5 @@
 def version() -> str:
-    import pkg_resources
+    import pkg_resources  # pylint: disable=import-outside-toplevel
     return pkg_resources.resource_string(__name__, 'hail_version').decode().strip()