+++ hail/python/hail/docs/change_log.md
@@ -22,6 +22,19 @@ Please note that **forward compatibility should not be expected, especially
 relating to file formats**: this means that it may not be possible to use
 an earlier version of Hail to read files written in a later version.
 
+## Version 0.2.53
+
+Released 2020-07-30
+
+### Bug fixes
+
+- (hail#9173) Use less confusing column key behavior in MT.show.
+- (hail#9172) Add a missing Python dependency to Hail: google-cloud-storage.
+- (hail#9170) Change Hail tree aggregate depth logic to correctly respect the
+  branching factor set in `hl.init`.
+
+---
+
 ## Version 0.2.52
 
 Released 2020-07-29
@@ -1117,4 +1130,4 @@ Released 2018-12-07
 
 We didn't start manually curating information about user-facing changes until version 0.2.4.
 
-The full commit history is available [here](https://github.com/hail-is/hail/commits/master).
+The full commit history is available [here](https://github.com/hail-is/hail/commits/main).